+++ tests/store/tracking/test_rest_store.py
@@ -833,80 +833,20 @@ def test_search_traces():
     assert token == "token"
 
 
-def test_search_traces_with_uc_schema_errors():
+def test_search_traces_errors():
     creds = MlflowHostCreds("https://hello")
     store = RestStore(lambda: creds)
     with pytest.raises(
         MlflowException,
-        match="Searching traces by UC schema is only supported by Databricks backend",
+        match="Searching traces by UC schema is not supported on the current tracking server",
     ):
         store.search_traces(uc_schemas=["catalog.schema"])
 
-
-def test_search_unified_traces():
-    """Test the search_traces method when using SearchUnifiedTraces with sql_warehouse_id."""
-    creds = MlflowHostCreds("https://hello")
-    store = RestStore(lambda: creds)
-    response = mock.MagicMock()
-    response.status_code = 200
-
-    # Format the response (using TraceInfo format for online path)
-    response.text = json.dumps(
-        {
-            "traces": [
-                {
-                    "request_id": "tr-1234",
-                    "experiment_id": "1234",
-                    "timestamp_ms": 123,
-                    "execution_time_ms": 456,
-                    "status": "OK",
-                    "tags": [
-                        {"key": "k", "value": "v"},
-                    ],
-                    "request_metadata": [
-                        {"key": "key", "value": "value"},
-                    ],
-                }
-            ],
-            "next_page_token": "token",
-        }
-    )
-
-    # Parameters for search_traces
-    experiment_ids = ["1234"]
-    filter_string = "status = 'OK'"
-    max_results = 10
-    order_by = ["timestamp_ms DESC"]
-    page_token = "12345abcde"
-    sql_warehouse_id = "warehouse123"
-    model_id = "model123"
-
-    with mock.patch("mlflow.utils.rest_utils.http_request", return_value=response) as mock_http:
-        trace_infos, token = store.search_traces(
-            experiment_ids=experiment_ids,
-            filter_string=filter_string,
-            max_results=max_results,
-            order_by=order_by,
-            page_token=page_token,
-            sql_warehouse_id=sql_warehouse_id,
-            model_id=model_id,
-        )
-
-        # Verify the correct endpoint was called
-        call_args = mock_http.call_args[1]
-        assert call_args["endpoint"] == "/api/2.0/mlflow/unified-traces"
-
-        # Verify the correct trace info objects were returned
-        assert len(trace_infos) == 1
-        assert isinstance(trace_infos[0], TraceInfo)
-        assert trace_infos[0].trace_id == "tr-1234"
-        assert trace_infos[0].experiment_id == "1234"
-        assert trace_infos[0].request_time == 123
-        # V3's state maps to V2's status
-        assert trace_infos[0].state == TraceStatus.OK.to_state()
-        assert trace_infos[0].tags == {"k": "v"}
-        assert trace_infos[0].trace_metadata == {"key": "value", "mlflow.trace_schema.version": "3"}
-        assert token == "token"
+    with pytest.raises(
+        MlflowException,
+        match="Searching traces by model_id is not supported on the current tracking server.",
+    ):
+        store.search_traces(model_id="model_id")
 
 
 def test_get_artifact_uri_for_trace_compatibility():
